+++ examples/basic.ml
@@ -2,4 +2,4 @@
 
 let () =
   let metadata = Metadata.parse_file "test.mp3" in
-  List.iter (fun (k,v) -> Printf.printf "- %s: %s\n" k v) metadata
+  List.iter (fun (k, v) -> Printf.printf "- %s: %s\n" k v) metadata