+++ src/metadata.ml
@@ -15,26 +15,26 @@ module Make (E : CharEncoding.T) = struct
   let recode = E.convert
 
   module ID3 = struct
-    let parse ?max_size f =
+    let parse f =
       let failure, v2 =
-        try (false, ID3v2.parse ~recode ?max_size f) with _ -> (true, [])
+        try (false, ID3v2.parse ~recode f) with _ -> (true, [])
       in
       let v1 =
         try
           Reader.reset f;
-          ID3v1.parse ~recode ?max_size f
+          ID3v1.parse ~recode f
         with _ -> if failure then raise Invalid else []
       in
       v2 @ v1
 
-    let parse_file = Reader.with_file parse
+    let parse_file ?custom_parser file =
+      Reader.with_file ?custom_parser parse file
   end
 
-  (** Return the first application which does not raise invalid. *)
-  let rec first_valid l ?(max_size : int option) file =
+  let rec first_valid l file =
     match l with
       | f :: l -> (
-          try f ?max_size file
+          try f file
           with Invalid ->
             Reader.reset file;
             first_valid l file)
@@ -43,19 +43,25 @@ module Make (E : CharEncoding.T) = struct
   module Audio = struct
     let parsers = [ID3.parse; OGG.parse; FLAC.parse]
     let parse = first_valid parsers
-    let parse_file = Reader.with_file parse
+
+    let parse_file ?custom_parser file =
+      Reader.with_file ?custom_parser parse file
   end
 
   module Image = struct
     let parsers = [JPEG.parse; PNG.parse]
     let parse = first_valid parsers
-    let parse_file = Reader.with_file parse
+
+    let parse_file ?custom_parser file =
+      Reader.with_file ?custom_parser parse file
   end
 
   module Video = struct
     let parsers = [AVI.parse; MP4.parse]
     let parse = first_valid parsers
-    let parse_file = Reader.with_file parse
+
+    let parse_file ?custom_parser file =
+      Reader.with_file ?custom_parser parse file
   end
 
   module Any = struct
@@ -64,11 +70,12 @@ module Make (E : CharEncoding.T) = struct
     (** Genering parsing of metadata. *)
     let parse = first_valid parsers
 
-    (** Parse the metadatas of a file. *)
-    let parse_file = Reader.with_file parse
+    let parse_file ?custom_parser file =
+      Reader.with_file ?custom_parser parse file
 
     (** Parse the metadatas of a string. *)
-    let parse_string = Reader.with_string parse
+    let parse_string ?custom_parser file =
+      Reader.with_string ?custom_parser parse file
   end
 
   include Any