+++ src/metadataPNG.ml
@@ -1,7 +1,7 @@
 open MetadataBase
 module R = Reader
 
-let parse ?max_size:_ f : metadata =
+let parse f : metadata =
   if R.read f 8 <> "\x89PNG\x0d\x0a\x1a\x0a" then raise Invalid;
   let _ = R.int32_be f in
   if R.read f 4 <> "IHDR" then raise Invalid;
@@ -15,4 +15,4 @@ let parse ?max_size:_ f : metadata =
     ("bit_depth", string_of_int bit_depth);
   ]
 
-let parse_file = R.with_file parse
+let parse_file ?custom_parser file = R.with_file ?custom_parser parse file